Moved common files to common. Rewired T2
