+++ demidenko.nikita/T2/parsers.hpp
@@ -2,13 +2,12 @@
 #define PARSERS_HPP
 
 #include <complex>
+#include <parserTemplate.hpp>
 #include <string>
 #include "keysEnum.hpp"
-#include "parserTemplate.hpp"
 
 namespace demidenko
 {
-  using DelimeterI = detail::ParserI< const char* >;
   using KeyI = detail::ParserI< KeysEnum& >;
   using ComplexI = detail::ParserI< std::complex< double >& >;
   using UnsignedLongLongI = detail::ParserI< unsigned long long& >;
