+++ demidenko.nikita/common/delimeter.hpp
@@ -1,0 +1,11 @@
+#ifndef DELIMETER_HPP
+#define DELIMETER_HPP
+
+#include "parserTemplate.hpp"
+
+namespace demidenko
+{
+  using DelimeterI = detail::ParserI< const char* >;
+}
+
+#endif
