+++ demidenko.nikita/common/parserTemplate.hpp
