+++ demidenko.nikita/common/streamGuard.cpp
