+++ demidenko.nikita/common/streamGuard.hpp
